requirements.txt
logging
datetime
typing
pathlib
pandas
numpy
pyscop
psycopg_binary
shiny
shinywidgets
folium
plotly